Extract global types to application
