+++ assets/src/domain/eventEditor/services/apollo/types.ts
@@ -1,23 +1,4 @@
-import { BaseType } from '@appServices/apollo/mutations';
-
-export type EntityId = string;
-export type EntityDbId = number;
-
-export interface Entity extends BaseType {
-	id: EntityId;
-	dbId: EntityDbId;
-	name: string;
-	__typename?: string;
-}
-
-export interface Trashable {
-	isTrashed: boolean;
-}
-
-export interface EntityEdge<E = Entity, ConnectionTypeName = string> {
-	nodes: E[];
-	__typename: ConnectionTypeName;
-}
+import { Entity, EntityEdge, Trashable } from '@appServices/apollo/types';
 
 export interface Event extends Entity {
 	desc: string;
